feat: Change weight type from int to double in WorkoutExerciseValueObject and WorkoutSessionExerciseDto for improved precision

+++ lib/valueObjects/trainingService/workout_exercise_value_object.dart
@@ -2,7 +2,7 @@ import 'package:shape_up_app/enums/trainingService/measure_unit.dart';
 
 class WorkoutExerciseValueObject {
   final String exerciseId;
-  int? weight;
+  double? weight;
   int? repetitions;
   MeasureUnit measureUnit;
